refactor: split project components into client files
